fix: postPerformanceForm 요청 응답 type 분리
